+++ VTK/Imaging/vtkImageAccumulate.cxx
@@ -137,7 +137,7 @@ void vtkImageAccumulateExecute(vtkImageAccumulate *self,
   T *tempPtr;
   int *outPtrC;
   int numC, outIdx, *outExtent, *outIncs;
-  float *origin, *spacing;
+  double *origin, *spacing;
   unsigned long count = 0;
   unsigned long target;
   double sumSqr[3], variance;
